Updated reflection

+++ utils/io_utils.py
@@ -10,5 +10,5 @@ def load_image(path):
     return image_rgb
 
 def save_image(image, path):
-    image_bgr = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
+    image_bgr = cv2.cvtColor(image.astype(np.uint8), cv2.COLOR_RGB2BGR)
     cv2.imwrite(str(path), image_bgr)
